Added rules for filtering unmatched blue tiles
The example file has been increased to 14 puzzles. All of them are
solved by the algorithm.

The trick was to add to the filtering step the elimination of the blue
tiles whose possible place was already taken by a red. The other way
around was already implemented, but polished. Now it calculates it based
on the solution so far, instead of the board.